move skill subcomponents into own folder
