+++ src/App.vue
@@ -1,10 +1,6 @@
 <script setup>
 import HelloWorld from './components/HelloWorld.vue'
 import AttributesView from './components/AttributesView.vue'
-import TalentsView from './components/SkillsView.vue'
-import CombatsSkillView from './components/CombatSkillsView.vue'
-import CoreStatsView from './components/CoreStatsView.vue'
-import SkillsFilter from './components/SkillsFilter.vue'
 import AllSkillsView from './components/AllSkillsView.vue'
 </script>
 
@@ -19,10 +15,6 @@ import AllSkillsView from './components/AllSkillsView.vue'
   <main class="splitSection">
     <AttributesView />
     <AllSkillsView />
-    <!-- <SkillsFilter />
-    <CoreStatsView />
-    <CombatsSkillView />
-    <TalentsView /> -->
   </main>
 </template>
