+++ src/components/skillComponents/CoreStatsView.vue
