+++ src/components/skillComponents/SkillsView.vue
